+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/Item/ColorDetails.cs
@@ -11,24 +11,24 @@ namespace ArgentPonyWarcraftClient
         /// Gets the red channel value for the color.
         /// </summary>
         [JsonPropertyName("r")]
-        public long Red { get; set; }
+        public int Red { get; set; }
 
         /// <summary>
         /// Gets the green channel value for the color.
         /// </summary>
         [JsonPropertyName("g")]
-        public long Green { get; set; }
+        public int Green { get; set; }
 
         /// <summary>
         /// Gets the blue channel value for the color.
         /// </summary>
         [JsonPropertyName("b")]
-        public long Blue { get; set; }
+        public int Blue { get; set; }
 
         /// <summary>
         /// Gets the alpha (opacity) channel value for the color.
         /// </summary>
         [JsonPropertyName("a")]
-        public long Alpha { get; set; }
+        public int Alpha { get; set; }
     }
 }
