+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/PlayableSpecialization/TalentTier.cs
@@ -11,7 +11,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the player level at which the talent tier unlocks.
         /// </summary>
         [JsonPropertyName("level")]
-        public long Level { get; set; }
+        public int Level { get; set; }
 
         /// <summary>
         /// Gets references to the talents for this tier.
@@ -23,6 +23,6 @@ namespace ArgentPonyWarcraftClient
         /// Gets the index of the talent tier.
         /// </summary>
         [JsonPropertyName("tier_index")]
-        public long TierIndex { get; set; }
+        public int TierIndex { get; set; }
     }
 }
