+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/PvpTier/PvpTier.cs
@@ -17,7 +17,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the ID of the PvP tier.
         /// </summary>
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Gets the name of the PvP tier.
@@ -29,13 +29,13 @@ namespace ArgentPonyWarcraftClient
         /// Gets the minimum rating for the PvP tier.
         /// </summary>
         [JsonPropertyName("min_rating")]
-        public long MinRating { get; set; }
+        public int MinRating { get; set; }
 
         /// <summary>
         /// Gets the maximum rating for the PvP tier.
         /// </summary>
         [JsonPropertyName("max_rating")]
-        public long MaxRating { get; set; }
+        public int MaxRating { get; set; }
 
         /// <summary>
         /// Gets the media associated with the PvP tier.
@@ -53,6 +53,6 @@ namespace ArgentPonyWarcraftClient
         /// Gets the rating type for the PvP tier.
         /// </summary>
         [JsonPropertyName("rating_type")]
-        public long RatingType { get; set; }
+        public int RatingType { get; set; }
     }
 }
