+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/Reputation/ReputationTier.cs
@@ -17,18 +17,18 @@ namespace ArgentPonyWarcraftClient
         /// Gets the minimum value of the reputation tier.
         /// </summary>
         [JsonPropertyName("min_value")]
-        public long MinValue { get; set; }
+        public int MinValue { get; set; }
 
         /// <summary>
         /// Gets the maximum value of the reputation tier.
         /// </summary>
         [JsonPropertyName("max_value")]
-        public long MaxValue { get; set; }
+        public int MaxValue { get; set; }
 
         /// <summary>
         /// Gets the ID of the reputation tier.
         /// </summary>
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
     }
 }
