+++ src/ArgentPonyWarcraftClient/Models/ProfileApi/AccountProfile/AccountCharacter.cs
@@ -29,7 +29,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the ID of the character.
         /// </summary>
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Gets a reference to the character's realm.
@@ -65,6 +65,6 @@ namespace ArgentPonyWarcraftClient
         /// Gets the level of the character.
         /// </summary>
         [JsonPropertyName("level")]
-        public long Level { get; set; }
+        public int Level { get; set; }
     }
 }
