+++ src/ArgentPonyWarcraftClient/Models/ProfileApi/CharacterSpecializations/PvpTalentSlot.cs
@@ -17,6 +17,6 @@ namespace ArgentPonyWarcraftClient
         /// Gets the PvP talent slot number.
         /// </summary>
         [JsonPropertyName("slot_number")]
-        public long SlotNumber { get; set; }
+        public int SlotNumber { get; set; }
     }
 }
