+++ src/ArgentPonyWarcraftClient/Models/ProfileApi/CharacterStatistics/CorruptionStatistics.cs
@@ -11,18 +11,18 @@ namespace ArgentPonyWarcraftClient
         /// Gets the corruption level.
         /// </summary>
         [JsonPropertyName("corruption")]
-        public long Corruption { get; set; }
+        public int Corruption { get; set; }
 
         /// <summary>
         /// Gets the corruption resistance.
         /// </summary>
         [JsonPropertyName("corruption_resistance")]
-        public long CorruptionResistance { get; set; }
+        public int CorruptionResistance { get; set; }
 
         /// <summary>
         /// Gets the effective corruption.
         /// </summary>
         [JsonPropertyName("effective_corruption")]
-        public long EffectiveCorruption { get; set; }
+        public int EffectiveCorruption { get; set; }
     }
 }
