+++ src/ArgentPonyWarcraftClient/Models/ProfileApi/CharacterStatistics/RatingStatistic.cs
@@ -11,7 +11,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the rating.
         /// </summary>
         [JsonPropertyName("rating")]
-        public long Rating { get; set; }
+        public int Rating { get; set; }
 
         /// <summary>
         /// Gets the rating bonus.
